apps/backend/pyproject.toml
[project]
name = "edupredict-backend"
version = "0.1.0"
description = "EduPredict FastAPI backend with RBAC, PostgreSQL, and ML inference."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  "fastapi>=0.115.6",
  "uvicorn[standard]>=0.32.1",
  "pydantic>=2.10.3",
  "pydantic-settings>=2.6.1",
  "email-validator>=2.2.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "SQLAlchemy[asyncio]>=2.0.36",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "python-multipart>=0.0.12",
  "orjson>=3.10.12",
  "httpx>=0.28.1",
  "numpy>=2.1.3",
  "pandas>=2.2.3",
  "scikit-learn>=1.5.2",
  "lightgbm>=4.5.0",
  "shap>=0.46.0",
  "joblib>=1.4.2"
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.24.0",
  "aiosqlite>=0.20.0",
  "ruff>=0.8.4",
  "mypy>=1.13.0",
  "types-passlib>=1.7.7.20240819"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
